aituning_service/pyproject.toml
[project]
name = "aituning-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.4",
]
